Fixed lint. Removed unnecessary complexity in layer's style generation
