+++ web/client/components/catalog/editor/AdvancedSettings/RasterAdvancedSettings.js
@@ -30,4 +30,4 @@ export default ({
             </Col >
         </FormGroup>
     </CommonAdvancedSettings>);
-}
+};
